feat(util): detect and extract phone numbers from image.
