+++ lib/esimslib/util/__init__.py
@@ -1,4 +1,4 @@
 """Util Module"""
 
 from esimslib.util.logger import logger
-from esimslib.util.qr_code_detector import QRCodeDetector
+from esimslib.util.qr_code_processor import QRCodeProcessor
