+++ lib/setup.cfg
@@ -18,6 +18,8 @@ install_requires =
     numpy
     requests
     pyzbar
+    pytesseract
+    cached_property
 
 [bdist_wheel]
 universal = true
